requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile '.\requirements.in'
#
about-time==4.2.1
    # via alive-progress
alive-progress==3.1.5
    # via -r .\requirements.in
altair==5.4.1
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   openai
asgiref==3.8.1
    # via django
attrs==24.2.0
    # via
    #   jsonschema
    #   referencing
blinker==1.8.2
    # via
    #   flask
    #   streamlit
cachetools==5.5.0
    # via streamlit
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   flask
    #   streamlit
    #   typer
colorama==0.4.6
    # via
    #   click
    #   tqdm
contourpy==1.3.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
dash==2.18.0
    # via -r .\requirements.in
dash-core-components==2.0.0
    # via dash
dash-html-components==2.0.0
    # via dash
dash-table==5.0.0
    # via dash
distro==1.9.0
    # via openai
django==5.1.1
    # via -r .\requirements.in
flask==3.0.3
    # via dash
fonttools==4.53.1
    # via matplotlib
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via streamlit
grapheme==0.6.0
    # via alive-progress
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.2
    # via openai
idna==3.8
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==8.5.0
    # via dash
itsdangerous==2.2.0
    # via flask
jinja2==3.1.4
    # via
    #   altair
    #   flask
    #   pydeck
jiter==0.5.0
    # via openai
jsonschema==4.23.0
    # via altair
jsonschema-specifications==2023.12.1
    # via jsonschema
kiwisolver==1.4.7
    # via matplotlib
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
matplotlib==3.9.2
    # via -r .\requirements.in
mdurl==0.1.2
    # via markdown-it-py
narwhals==1.7.0
    # via altair
nest-asyncio==1.6.0
    # via dash
numpy==2.1.1
    # via
    #   contourpy
    #   matplotlib
    #   pandas
    #   pyarrow
    #   pydeck
    #   streamlit
openai==1.44.1
    # via -r .\requirements.in
packaging==24.1
    # via
    #   altair
    #   matplotlib
    #   plotly
    #   streamlit
pandas==2.2.2
    # via
    #   -r .\requirements.in
    #   streamlit
pillow==10.4.0
    # via
    #   -r .\requirements.in
    #   matplotlib
    #   streamlit
plotly==5.24.0
    # via dash
protobuf==5.28.1
    # via streamlit
pyarrow==17.0.0
    # via streamlit
pydantic==2.9.1
    # via openai
pydantic-core==2.23.3
    # via pydantic
pydeck==0.9.1
    # via streamlit
pygments==2.18.0
    # via rich
pyparsing==3.1.4
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.0.1
    # via -r .\requirements.in
pytz==2024.2
    # via pandas
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   -r .\requirements.in
    #   dash
    #   streamlit
retrying==1.3.4
    # via dash
rich==13.8.1
    # via
    #   streamlit
    #   typer
rpds-py==0.20.0
    # via
    #   jsonschema
    #   referencing
shellingham==1.5.4
    # via typer
six==1.16.0
    # via
    #   python-dateutil
    #   retrying
smmap==5.0.1
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
sqlparse==0.5.1
    # via django
st-pages==1.0.1
    # via -r .\requirements.in
streamlit==1.38.0
    # via
    #   -r .\requirements.in
    #   st-pages
tenacity==8.5.0
    # via
    #   plotly
    #   streamlit
toml==0.10.2
    # via streamlit
tornado==6.4.1
    # via streamlit
tqdm==4.66.5
    # via openai
typer==0.12.5
    # via -r .\requirements.in
typing-extensions==4.12.2
    # via
    #   altair
    #   dash
    #   openai
    #   pydantic
    #   pydantic-core
    #   streamlit
    #   typer
tzdata==2024.1
    # via
    #   django
    #   pandas
urllib3==2.2.2
    # via requests
watchdog==4.0.2
    # via streamlit
werkzeug==3.0.4
    # via
    #   dash
    #   flask
zipp==3.20.1
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
